% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_differential_analysis.R
\name{AdjustDESeqContrasts}
\alias{AdjustDESeqContrasts}
\title{Title}
\usage{
AdjustDESeqContrasts(model.contrasts, padj.alpha = 0.1, theta = seq(from =
  0, to = 0.5, by = 0.01), pAdjustMethod = "BH", returnMeta = TRUE)
}
\arguments{
\item{padj.alpha}{}

\item{theta}{}

\item{pAdjustMethod}{}

\item{returnMeta}{}

\item{dds.contrasts}{}
}
\description{
Title
}
