% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{AnnotateCounts}
\alias{AnnotateCounts}
\title{Annotate count data with AnnotationDbi}
\usage{
AnnotateCounts(counts, species, key.type, column.keys = c("SYMBOL",
  "ENTREZID", "ENSEMBL", "GENENAME"), multiVals = "first")
}
\arguments{
\item{counts}{The count data as read with \code{\link{ReadCounts()}}.}

\item{species}{The species used for the gene annotation}

\item{key.type}{Keytype used for gene annotation, e.g. ENSEMBL, Entrez, etc.}

\item{column.keys}{A named vector of other keytypes to add to the annotation using the \pkg{AnnotationDbi} package}

\item{multiVals}{What to do with multiple matching annotations , passed to \code{\link{mapIds}}}
}
\value{
A data.frame with all chosen annotations according to \code{column.keys}
}
\description{
Annotate count data with AnnotationDbi
}
