% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_gene_filter.R
\name{deseq_gene_filter}
\alias{deseq_gene_filter}
\title{deseq_gene_filter}
\usage{
deseq_gene_filter(dds, flist, normalized = FALSE)
}
\arguments{
\item{dds}{A DESeqDataset object, easily made using \code{\link{DESeqDataSetFromMatrix()}}}

\item{flist}{A list of filter functions to be applied to the array.
Each function should return either TRUE or FALSE for every gene. Genes for which
the function returns FALSE are removed.}
}
\description{
A helper function for removing genes according to various criteria.
}
