% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_sample_subset.R
\name{deseq_sample_subset}
\alias{deseq_sample_subset}
\title{Subset deseq dataset by samples}
\usage{
deseq_sample_subset(dds, subset.list, verbose = FALSE)
}
\arguments{
\item{dds}{A DESeqDataSet object with the sample data in colData(dds)}

\item{subset.list}{A list of two-element vectors. The first element of every vector is the column / feature name, all further elements are feature levels to keep.}

\item{verbose}{Logical. Whether to output a message listing the removed samples (default FALSE)}
}
\description{
Subset deseq dataset by samples
}
