% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_transform.R
\name{deseq_transform}
\alias{deseq_transform}
\title{deseq_transform}
\usage{
deseq_transform(dds, blind = TRUE, method = "scale")
}
\arguments{
\item{dds}{A DESeqDataSet object.}
}
\value{
A data.frame containing the transformed counts.
}
\description{
Perform deseq2 transformations such as rlog and vst. Returns a data.frame contaned the transformed counts.
}
