% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_differential_analysis.R
\name{DESeqAnalysis}
\alias{DESeqAnalysis}
\title{Title}
\usage{
DESeqAnalysis(dds, comparisonFile, annotationData = NULL)
}
\arguments{
\item{dds}{}

\item{comparisonFile}{}

\item{annotationData}{}
}
\description{
Title
}
