% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_filter_mean.R
\name{gene_filter_mean}
\alias{gene_filter_mean}
\title{Filter genes by mean expression}
\usage{
gene_filter_mean(x, cutoff)
}
\arguments{
\item{x}{}

\item{cutoff}{}
}
\description{
To be used as an argument to \code{\link{genefilter}} from the \pkg{genefilter} package.
}
