% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_filter_prevalence.R
\name{gene_filter_prevalence}
\alias{gene_filter_prevalence}
\title{gene_filter_prevalence}
\usage{
gene_filter_prevalence(x, cutoff = 5, threshold = 0.05)
}
\description{
Function for filtering genes. Given a gene array function returns TRUE if gene counts
occur less than the cutoff values for at least as the threshold percentage of samples.
}
