% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{getData}
\alias{getData}
\title{Title}
\usage{
getData(countdata, coldata, species = NULL, keytype = NULL,
  annotationdata = NULL, calc.sizeFactors = TRUE, calc.disps = TRUE,
  calc.vst = FALSE)
}
\arguments{
\item{countdata}{Filename of count / expression data in a format where rownames are genes/features and columns are samples.}

\item{coldata}{Filename of column / sample data. Sample data has samples in the rows and features in the columns. The rows of colData should preferably equal the column names of the count data in the same order.}

\item{species}{The species of the gene annotations, e.g. Mouse, Human, Rat, etc.}

\item{annotationdata}{(OPTIONAL) Filename of annotation data with gene names in the rownames, equal to the count data, and different annotations in the columns. E.g. ENSEMBL gene id's in the rownames, and SYMBOL and Entrez Id's in other columns.}

\item{key.type}{The gene id key, e.g. ENSEMBL, ENTREZ or SYMBOL etc.}
}
\description{
Title
}
