% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_differential_analysis.R
\name{GetDESeqContrasts}
\alias{GetDESeqContrasts}
\title{Title}
\usage{
GetDESeqContrasts(dds.model, contrasts, model.index = 1, padj.alpha = 0.1)
}
\arguments{
\item{contrast}{}
}
\description{
Title
}
