% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{LoadAnnotationDb}
\alias{LoadAnnotationDb}
\title{Load the correct annotation database given the species}
\usage{
LoadAnnotationDb(gene.species)
}
\arguments{
\item{gene.species}{}
}
\description{
Load the correct annotation database given the species
}
