% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{ParseComparisons}
\alias{ParseComparisons}
\title{Parse rna-seq comparison file}
\usage{
ParseComparisons(compData)
}
\arguments{
\item{compData}{Comparison data in data.frame from \code{\link{ReadComparisons}}}
}
\value{
list of designs and contrasts
}
\description{
Parse rna-seq comparison file
}
