% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMDS.R, R/plot_deseq_MDS.R
\name{plot_deseq_mds}
\alias{plot_deseq_mds}
\alias{plot_deseq_mds}
\title{Glimma MDS plot from DESeq2 variance-stabilized counts}
\usage{
plot_deseq_mds(vsd, groups.vec, folder = "Glimma", main = NULL,
  html = "MDS-Plot", launch = TRUE)

plot_deseq_mds(vsd, groups.vec, folder = "Glimma", main = NULL,
  html = "MDS-Plot", launch = TRUE)
}
\arguments{
\item{vsd}{Variance stabilized counts. Recommended to put in mcols(dds)$vst}

\item{groups.vec}{Group names to color by. should be column names of colData(dds)}

\item{folder}{Output folder of .html output, either relative or absolute}

\item{main}{Title of MDS plot}

\item{html}{Output filename of .html file}

\item{launch}{Whether to launch a browser to view the Glimma MDS plot}

\item{vsd}{Variance stabilized counts. Recommended to put in mcols(dds)$vst}

\item{groups.vec}{Group names to color by. should be column names of colData(dds)}

\item{folder}{Output folder of .html output, either relative or absolute}

\item{main}{Title of MDS plot}

\item{html}{Output filename of .html file}

\item{launch}{Whether to launch a browser to view the Glimma MDS plot}
}
\value{



}
\description{
Glimma MDS plot from DESeq2 variance-stabilized counts

Glimma MDS plot from DESeq2 variance-stabilized counts
}
\details{
The \code{vsd} is expected to be DESeqTransform object, output from \code{\link{VarianceStabilizingTransformation}}.

The \code{vsd} is expected to be DESeqTransform object, output from \code{\link{VarianceStabilizingTransformation}}.
}
