% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_pairs.R
\name{plot_gene_pairs}
\alias{plot_gene_pairs}
\title{Pairplot of genes and selected features}
\usage{
plot_gene_pairs(dds, features = NULL, colour = NULL)
}
\arguments{
\item{dds}{A heavily filtered DESeqDataSet, i.e. a maximum of 10 genes}

\item{features}{Names of features to include in the pairplot so they can be correlated with gene expression. Need to correspond to column names in colData(dds).}

\item{colour}{Single feature name to colour plots by.}
}
\description{
Pairplot of genes and selected features
}
