% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_differential_plotting.R
\name{PlotDESeqAnalysis}
\alias{PlotDESeqAnalysis}
\title{Title}
\usage{
PlotDESeqAnalysis(dds.analysis, annotationData = NULL, groups = NULL,
  color.var = NULL, alpha = 0.05, normalized = TRUE, glimma.alpha = 1,
  folder = "./Glimma", main = NULL, html = "MD-Plot", verbose = 1, ...)
}
\arguments{
\item{dds.analysis}{}

\item{annotationData}{}

\item{groups}{}

\item{color.var}{}

\item{alpha}{}

\item{normalized}{}

\item{glimma.alpha}{}

\item{folder}{}

\item{main}{}

\item{html}{}

\item{verbose}{}

\item{...}{}
}
\description{
Title
}
