% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_differential_plotting.R
\name{PlotGlimmaContrast}
\alias{PlotGlimmaContrast}
\title{Plot Glimma MD plot for DESeq2 results}
\usage{
PlotGlimmaContrast(dds, dds.res, annotationData, groups = NULL,
  color.var = NULL, contrast = NULL, alpha = 0.05, normalized = TRUE,
  glimma.alpha = 1, folder = "./Glimma", main = NULL, html = "MD-Plot",
  launch = FALSE, ...)
}
\arguments{
\item{dds}{DESeq dataset with computed sizeFactors and dispersions}

\item{dds.res}{DESeqResults object - or alternatively only the results in a data.frame}

\item{annotationData}{Annotation data as data.frame}

\item{groups}{A single value or vector of character values of features in the colData of dds to separate groups by in the plot}

\item{color.var}{Sample variable/feature to color expression plot by}

\item{contrast}{Contrast vector used in DESeq's \code{\link{results}} function. 
Used for determining groups in the expression plot only if neither \code{groups} or \code{color.var} is not given}

\item{alpha}{P-adjusted significance level for coloring DEG's}

\item{normalized}{Logical whether to use normalized expression values. Default TRUE}

\item{glimma.alpha}{P-value cutoff to eliminate genes from glimma plot. May save diskspace by removing uninteresting genes. Default is 1 (does not remove anything)}

\item{folder}{Output folder. Default to /Glimma subdirectory}

\item{main}{Title of the glimma plot}

\item{html}{HTML file output title}

\item{launch}{Boolean to open Glimma plot output in browser. Default FALSE (use False in reports!)}
}
\value{
Saves a Glimma plot (.html file) to the specified directory. Returns relative file location as a string.
}
\description{
Helper function for DESeq2 analysis reports
Plot an MD plot of DESeq2 results with a specific contrast
- Requires the Glimma package
Outputs a Glimma MD plot .html file into a specified subfolder of the report
}
