% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_differential_analysis.R
\name{pvalueAdjustment}
\alias{pvalueAdjustment}
\title{Adjust p-values of DESeq2 results in data.frames}
\usage{
pvalueAdjustment(res, filter, theta, alpha = 0.1, pAdjustMethod = "BH",
  independentFiltering = TRUE)
}
\arguments{
\item{res}{A data.frame or DESeqResults object}

\item{filter}{defaults to the baseMean column of the results. Otherwise takes a vector of the same lengt has the number of rows of \code{res}}

\item{independentFiltering}{Logical whether to apply independent filtering. Defaults to TRUE}
}
\description{
Adjust p-values using independentFiltering the way DESeq2 does it when using the \code{\link[DESeq2]{results}} function.
The only difference is that this function works with data.frames as well, and therefore does not update the metadata.
Only DESeqResults objects have a metadata method.
}
\details{
Copied from
\url{https://github.com/mikelove/DESeq2/blob/master/R/results.R}

Copied because it is not exported and we can use it to do independentFiltering of
combined results in data.frames instead of DESeqResults objects

Does not adjust metadata of results, because there isn't any.
}
