% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{ReadAnnotation}
\alias{ReadAnnotation}
\title{Read rna-seq annotation data}
\usage{
ReadAnnotation(ANNOTATIONFILE, sep = NULL, quote = "\\"'")
}
\arguments{
\item{ANNOTATIONFILE}{Path to annotation file. File should have gene ID's in the rows, same gene ID's as the expression data.}

\item{sep}{(Optional) Separator of file, determined automatically if not given}

\item{quote}{parameter passed to \code{\link{read.table}}. Default disables quoting}
}
\value{
data frame of annotation data
}
\description{
Read rna-seq annotation data
}
