% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{ReadColData}
\alias{ReadColData}
\title{Read rna-seq coldata file (sample data)}
\usage{
ReadColData(COLDATAFILE, sep = NULL, sort.samples = TRUE)
}
\arguments{
\item{COLDATAFILE}{Path to annotation file. File should have gene ID's in the rows, same gene ID's as the expression data.}

\item{sep}{(Optional) Separator of file, determined automatically if not given}

\item{sort.samples}{Logical whether to sort samples by rownames. Default TRUE. This will make is easier to align expression data columns and sample data rownames.}
}
\value{
A data.frame of the sample data
}
\description{
This function is used to read coldata files, or sample data files. Every row is a sample and every column a feature.
The rownames should be the same as the column names in the expression data. Or, at least the expression columnnames and sample data rownames should overlap.
By default the sample and expression data are sorted by their rownames so they automatically align if all sample names overlap.
}
