% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{ReadComparisons}
\alias{ReadComparisons}
\title{Read comparison file}
\usage{
ReadComparisons(COMPARISONFILE, sep = NULL)
}
\arguments{
\item{COMPARISONFILE}{}

\item{sep}{}
}
\description{
Read comparison file
}
