% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{ReadCounts}
\alias{ReadCounts}
\title{Read rna-seq count data}
\usage{
ReadCounts(COUNTFILE, sep = NULL, sort.samples = TRUE)
}
\arguments{
\item{COUNTFILE}{}

\item{sep}{}

\item{sort.samples}{}
}
\description{
Read rna-seq count data
}
