% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_deseq_results.R
\name{SummariseModelResults}
\alias{SummariseModelResults}
\title{Summarise deseq2 results}
\usage{
SummariseModelResults(dds.analysis, model.index = NULL)
}
\arguments{
\item{dds.analysis}{Data.frame of DESeq2 results, with the same kind of format as the output of \code{\link{results}}.}

\item{model.index}{If not \code{NULL} will only print combined summary of 1 model}
}
\value{
The output is a printed summary of the results
}
\description{
This function makes the same type of summary output as the \code{\link{summary}} function from \pkg{DESeq2}.
The difference being that this function will also work on data.frame objects, especially handy when we combine
DESeqResults objects together in a long-format data.frame. The code was largely copied from \code{\link{summary}}
with some minor adjustments.
}
\details{
The \code{\link{summary}} function from DESeq2 will only work with DESeqResults objects in R
}
