% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_SubsetFunctions.R,
%   R/WGCNA_ThresholdSubsets.R
\name{WGCNA_PlotThresholds}
\alias{WGCNA_PlotThresholds}
\alias{WGCNA_PlotThresholds}
\title{Plot soft-threshold estimations from subsets}
\usage{
WGCNA_PlotThresholds(fitIndices, powerEstimates, RsquaredCut = 0.85)

WGCNA_PlotThresholds(fitIndices, powerEstimates, RsquaredCut = 0.85)
}
\arguments{
\item{fitIndices}{A data.frame of fitIndices including a column \code{SignedR2} and \code{Subset}, as output from \code{\link{WGCNA_ThresholdSubsets}}}

\item{powerEstimates}{A named list of powerEstimates for all subsets.}

\item{RsquaredCut}{Cutoff of the minimal r-squared needed to accept soft-threshold power.}

\item{fitIndices}{A data.frame of fitIndices including a column \code{SignedR2} and \code{Subset}, as output from \code{\link{WGCNA_ThresholdSubsets}}}

\item{powerEstimates}{A named list of powerEstimates for all subsets.}

\item{RsquaredCut}{Cutoff of the minimal r-squared needed to accept soft-threshold power.}
}
\value{
A plotly subplot of the signed R^2 and mean network connectivity versus the soft-threshold.

A plotly subplot of the signed R^2 and mean network connectivity versus the soft-threshold.
}
\description{
Uses the output from \code{\link{WGCNA_ThresholdSubsets}} and plots the results in an interactive plot of the signed R-squared and the mean network connectivity.

Uses the output from \code{\link{WGCNA_ThresholdSubsets}} and plots the results in an interactive plot of the signed R-squared and the mean network connectivity.
}
