% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_SubsetFunctions.R
\name{WGCNA_SubsetModules}
\alias{WGCNA_SubsetModules}
\title{BlockWise Modules of subsets}
\usage{
WGCNA_SubsetModules(datExpr, subsets, power, include.total = TRUE, ...)
}
\arguments{
\item{datExpr}{}

\item{subsets}{}

\item{power}{}

\item{include.total}{}

\item{...}{passed to blockwiseModules()}
}
\description{
BlockWise Modules of subsets
}
