% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_SubsetFunctions.R,
%   R/WGCNA_ThresholdSubsets.R
\name{WGCNA_ThresholdSubsets}
\alias{WGCNA_ThresholdSubsets}
\alias{WGCNA_ThresholdSubsets}
\title{Determine WGCNA soft threshold with DESeqDataSet}
\usage{
WGCNA_ThresholdSubsets(datExpr, subsets = NULL, powers = c(1:10, seq(12, 30,
  2)), ...)

WGCNA_ThresholdSubsets(datExpr, subsets = NULL, powers = c(1:10, seq(12, 30,
  2)), ...)
}
\arguments{
\item{datExpr}{A matrix or data.frame with samples in rows and gene expressions in columns. Recommended to use variance-stabilized counts}

\item{subsets}{A list of integer vectors to determine different subsets. 
Preferably a named list where names indicate a short name to differentiate subsets. Default is NULL, in which case we use all data.}

\item{powers}{Numeric vector of powers (soft-thresholds) to test}

\item{...}{Additional arguments passed to \code{\link{pickSoftThreshold}} (TODO)}

\item{datExpr}{A matrix or data.frame with samples in rows and gene expressions in columns. Recommended to use variance-stabilized counts}

\item{subsets}{A list of integer vectors to determine different subsets. 
Preferably a named list where names indicate a short name to differentiate subsets. Default is NULL, in which case we use all data.}

\item{powers}{Numeric vector of powers (soft-thresholds) to test}

\item{...}{Additional arguments passed to \code{\link{pickSoftThreshold}} (TODO)}
}
\value{
A two-element list. First the \code{powerEstimates}, the determined powers for each subset. 
     Second the \code{fitIndices}, the combined individual data.frame outputs from \code{\link{pickSoftThreshold}}.

A two-element list. First the \code{powerEstimates}, the determined powers for each subset. 
     Second the \code{fitIndices}, the combined individual data.frame outputs from \code{\link{pickSoftThreshold}}.
}
\description{
Attempt to determine soft-threshold for WGCNA similarity measures. Effectively we raise the absolute correlations between gene expressions to a power,
and we aim to determine the minimal power required to achieve an approximately scale-free network topology.

Attempt to determine soft-threshold for WGCNA similarity measures. Effectively we raise the absolute correlations between gene expressions to a power,
and we aim to determine the minimal power required to achieve an approximately scale-free network topology.
}
\details{
Calls the function \code{\link{pickSoftThreshold}} from the \pkg{WGCNA} package using biweight midcorrelation (bicor) and a signed network by default.

You may use this function especially for differential network analysis, where one needs to compute weighted networks based on subsets of the data.
This function may help to determine the best soft-threshold, which needs to be determined for both subsets individually.

Calls the function \code{\link{pickSoftThreshold}} from the \pkg{WGCNA} package using biweight midcorrelation (bicor) and a signed network by default.

You may use this function especially for differential network analysis, where one needs to compute weighted networks based on subsets of the data.
This function may help to determine the best soft-threshold, which needs to be determined for both subsets individually.
}
