% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_functions.R
\name{ReadComparisons}
\alias{ReadComparisons}
\title{Read comparison file}
\usage{
ReadComparisons(COMPARISONFILE, sep = NULL)
}
\arguments{
\item{COMPARISONFILE}{Path to comparison file in .csv format(!)}

\item{sep}{(Optional) Separator of file. Determined automatically if not given.}
}
\description{
Read comparison file
}
